Refactor logging to use context-aware methods
Replaced `slog` calls with their context-aware counterparts (`slog.*Context`) to improve log traceability and debugging in context-dependent operations. Additionally, removed redundant log fields and unused code for cleaner implementation.
