+++ pkg/bot/handlers.go
@@ -26,10 +26,7 @@ func (s *ServiceImpl) setupHandler() Handler {
 }
 
 func (s *ServiceImpl) handleMessage(ctx context.Context, message *tgbotapi.Message) (tgbotapi.MessageConfig, error) {
-	slog.Info("Received message",
-		slog.Int64("chat_id", message.Chat.ID),
-		slog.String("text", message.Text),
-	)
+	slog.DebugContext(ctx, "Received message", slog.String("text", message.Text))
 
 	if message.Text == "" {
 		return tgbotapi.MessageConfig{}, nil
